Moved method str_to_bool closer to its use in namespace app.schema
