+++ source/app/util.py
@@ -136,16 +136,3 @@ def hmac_verify(signature_enc, data):
         return True
     except InvalidSignature:
         return False
-
-
-def str_to_bool(value):
-    if value is None:
-        return False
-
-    if isinstance(value, bool):
-        return value
-
-    if isinstance(value, int):
-        return bool(value)
-
-    return value.lower() in ['true', '1', 'yes', 'y', 't']
